+++ src_vue/config/config.json
@@ -2,31 +2,28 @@
   "flask_host": "0.0.0.0",
 
   "collections_info_backend": {
-    "default": "mysql://user:password@db/scot",
-    "en_books": "mysql://user:password@db/scot",
-    "fi_news": "mysql://scot:scot@ltdatabase1/DT_finnews_dependency",
-    "de70": "mysql://scot:scot@ltdatabase1/DT_de70mlcc_dependency"
+    "default": "mysql://root:root@127.0.0.1:3377/scot",
+    "en_books": "mysql://root:root@127.0.0.1:3377/scot",
+    "fi_news": "mysql://root:root@127.0.0.1:3306/DT_finnews_dependency",
+    "de70": "mysql://root:root@127.0.0.1:3306/DT_de70mlcc_dependency"
   },
 
   "collections_info_elastic": {
     "fi_news": {
       "es_index": "fi_news",
-      "es_host": "elasticsearch",
-      "es_port": 9200
+      "es_host": "127.0.0.1",
+      "es_port": 9292
     },
-    "de70": {
-      "es_index": "de70",
-      "es_host": "elasticsearch",
-      "es_port": 9200
-    }
+    "de70": { "es_index": "de70", "es_host": "127.0.0.1", "es_port": 9292 }
   },
+
   "collections_info_frontend": {
     "English Books": {
       "key": "en_books",
       "target": "crisis/NN",
       "p": 100,
       "d": 20,
-      "info": "English Google Books, 1520-2008."
+      "info": "English Google Books, 1520-2008"
     },
     "German News 70m Dep": {
       "key": "de70",
@@ -40,7 +37,7 @@
       "target": "Yksi#NUM",
       "p": 100,
       "d": 8,
-      "info": "Various Finnish Online News - 14 Million Sentences, "
+      "info": "Various Finnish Online News - 14 Million Sentences"
     }
   }
 }
