+++ src_vue/model/orm_models.py
@@ -3,12 +3,11 @@ from typing import List, Dict
 from flask_sqlalchemy import SQLAlchemy
 
 """  
+NOT IMPLEMENTED BUT STILL HELFPUL TO UNDERSTAND THE DB-COMPONENT
 This dataclass defines SQL-ORM for the raw Graph, ie the one consisting of directed and not overlaid edges and node-information 
 [A node in the network is really an edge between the target-word and the node]
 If further defines the ORM for features-data
-TODO: Migrate ORM to sqlalchemy later but keep records at the moment
 """
-
 
 
 # ----------- ORMS
@@ -20,6 +19,7 @@ class RawLink:
     time_id: int = None
     score: float = None
     feature: str = None
+
 
 @dataclass
 class RawFeature:
