+++ src_vue/persistence/documentdb.py
@@ -6,7 +6,7 @@ import json
 class Documentdb:
 
     def __init__(self, el_host, el_port):
-        
+
         try:
             self.es = Elasticsearch([{'host': el_host, 'port': el_port}])
         except:
@@ -14,24 +14,24 @@ class Documentdb:
 
     def search(self, jo, bim, es_index="corona_news"):
         self.es.indices.refresh(index=es_index)
-        res = self.es.search(index=es_index, 
-            body={
-                "query": {
-      	            "nested":{
-      		            "path":"jobim",
-      		                "query": {
-        		                "bool": {
-        			                "must": [
-            			                { "match": { "jobim.jo": jo }},
-            			                { "match": { "jobim.bim": bim }}
-            		                ]
-        		                }
-    			            }
-    		        }
-		        }
+        res = self.es.search(index=es_index,
+                             body={
+                                 "query": {
+                                     "nested": {
+                                         "path": "jobim",
+                                         "query": {
+                                             "bool": {
+                                                 "must": [
+                                                     {"match": {"jobim.jo": jo}},
+                                                     {"match": {"jobim.bim": bim}}
+                                                 ]
+                                             }
+                                         }
+                                     }
+                                 }
 
-            }
+                             }
 
-        )
-        #print(res)
+                             )
+        # print(res)
         return res
