--- src_vue/services/feature_information.py
@@ -1,3 +1,0 @@
-from persistence.documentdb import Documentdb
-from persistence.db import Database
-
