+++ src_vue/static/.prettierrc
@@ -1,4 +1,4 @@
 {
-  "tabWidth": 4,
+  "tabWidth": 2,
   "useTabs": false
 }
