+++ src_vue/templates/index.html
@@ -17,7 +17,8 @@
 
 	<!-- Other JavaScript files -->
 	<script type="text/javascript" src=".%% url_for('static', filename='render_sense_graph.js') %%"></script>
-
+	<script type="text/javascript" src=".%% url_for('static', filename='data.js') %%"></script>
+	<script type="text/javascript" src=".%% url_for('static', filename='inout.js') %%"></script>
 	<!-- Stylesheets -->
 	<link type="text/css" rel="stylesheet" href="https://unpkg.com/bootstrap/dist/css/bootstrap.min.css" />
 	<link type="text/css" rel="stylesheet" href="https://unpkg.com/bootstrap-vue@2.11.0/dist/bootstrap-vue.css" />
